refactor: menu group 등록 api에서 request로 dto를 사용하도록 수정
