+++ README.md
@@ -122,18 +122,19 @@
 ## 리팩터링
 
 - 테스트
-    - [ ] nested 어노테이션 사용해보기
-    - [ ] 인수 테스트도 추가해보기
-    - [ ] repository, dao 테스트
+    - nested 어노테이션 사용해보기
+    - 인수 테스트도 추가해보기
+    - repository, dao 테스트
 - 도메인
-    - [ ] 불필요한 setter 없애기
-    - [ ] 도메인에서 처리 가능한 비지니스 로직을 서비스로부터 책임 분리
-    - [ ] 상황에 따라 필요한 부생성자 만들기
-    - [ ] DB에 저장하기 위해 형태가 다른 생성자를 사용하는 문제를 dto로 해결할 방법 고민해보기
+    - 불필요한 setter 없애기
+    - 도메인에서 처리 가능한 비지니스 로직을 서비스로부터 책임 분리
+    - 상황에 따라 필요한 부생성자 만들기
+    - DB에 저장하기 위해 형태가 다른 생성자를 사용하는 문제를 dto로 해결할 방법 고민해보기
 - 컨트롤러
-    - [ ] request를 domain 대신 dto로 받도록 수정
+    - request를 domain 대신 dto로 받도록 수정
 - repository & service
     - 조회시 존재하지 않는 경우 예외 처리를 어디서 하면 좋을지 고민해보기
+    - transaction을 적절하게 수정해보기
 - dao
     - dao에 @Repository를 사용할지 @Component를 사용할지 생각히보기
 - etc
