+++ src/main/java/kitchenpos/application/MenuGroupService.java
@@ -3,6 +3,7 @@ package kitchenpos.application;
 import java.util.List;
 import kitchenpos.dao.MenuGroupDao;
 import kitchenpos.domain.MenuGroup;
+import kitchenpos.ui.request.menugroup.MenuGroupCreateRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,7 +17,7 @@ public class MenuGroupService {
     }
 
     @Transactional
-    public MenuGroup create(final MenuGroup request) {
+    public MenuGroup create(final MenuGroupCreateRequest request) {
         return menuGroupDao.save(new MenuGroup(request.getName()));
     }
