+++ src/main/java/kitchenpos/ui/request/menugroup/MenuGroupCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.request.menu;
+package kitchenpos.ui.request.menugroup;
 
 public class MenuGroupCreateRequest {
